move telnet to own file
